Исправление замечаний ДЗ-1
