+++ src/human/Sex.java
@@ -1,3 +1,5 @@
+package human;
+
 public enum Sex {
     MALE,
     FEMALE
